Working 2D steady state test, but model fails it
The problem is the open boundary conditions, which do not mimic to Dirichlet
boundaries requied by the exact solution. The intial height surface has a step
at two corners of the grid, and this propagates along the edges (as it should
for open boundaries).

There are two remedies:

1. Implement a more flexible model that can handle constant boundary
conditions (and closed boundaries as well)

This can probably be done elegantly by dynamically selecting the slice range
when computing 2nd derivative terms. However, I am not interested in Dirichlet
boundaries for the experiments I have planned, and I still would not have a
test for the open boundary condition case.

2. Pick another test case that does work for open boundaries.

This is prefered solution, since I need a test for the open boundary case no
matter what.