New parameterization for hillslope kappa in ice-cascade
The change solves an upcoming problem - after turning diffusion on/off due to
ice or water cover, I need to know what value to use when I turn it on again. A
simple solution is to ask the user for two value for kappa: a value for
"active" diffusion, and a value for "inactive" diffusion under ice/water. The
user supplies two scalars, and the grid is populated from these.

Note that the hillslope model remains more general, but the core ice-cascade
model only uses the features it needs.