Trimmed fat from ice-cascade core
+ Removed CLI interface
+ Remove start-from-file and write-input-file functionality
+ Removed option to clobber output files
+ Removed interactive display option (just view results)

The new simpler approach is for user to initialize and run models in the python
environment. This makes sense considering that there is no other practical way
to initialize the model --- even using an input file, the user is likely to
generate it in Python (or similar). Having a single method for init and run
greatly simplifies my code.